requirements_render.txt
Flask==2.3.3
alpaca-py==0.20.0
pandas==2.0.3
plotly==5.15.0
python-dotenv==1.0.0
gunicorn==21.2.0
